kubernetes network policy enhancements
Kubernetes network policy spec states that “multiple network policies
can select the same set of pods. In this case, the ingress rules for
each are combined additively.”. Thus, it was necessary to add a new set
of policy rules structures, kubernetes specific, that use the
LabelSelector from kubernetes to check if a particular set of labels is
covered by that kubernetes rule.

Also, since the same set of rules should be used additively, when the
kubernetes network policy is parsed to a cilium network policy, the
default action was changed from ALWAYS_ACCEPT to ACCEPT. This will force
the SearchContext to go through all the cilium kubernetes tree and make
sure there isn’t a single rule denying that SearchContext.

Since the NamespaceSelector matches on kubernetes namespaces’ labels,
the kuberentes network policies will be parsed and have all the keys of
the NamespaceSelector prefixed with `cilium.k8s.ns-labels` to store all
labels from the namespaces in the cilium policy.

Signed-off-by: André Martins <andre@cilium.io>